requirements.txt
pandas==2.2.1
xarray==2022.12.0
pv-site-prediction==0.1.19
pydantic==2.6.2
huggingface_hub== 0.17.3 # only for evaluation
python-dotenv==1.0.1
openmeteo-requests==1.2.0
requests-cache==1.2.0
retry-requests==2.0.0
gdown==5.1.0
xgboost==2.0.3
plotly
typer
streamlit
async_timeout
uvicorn
fastapi